School Projects/PadelConcept/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
asgiref==3.8.1
    # via
    #   django
click==8.1.7
    # via djlint
colorama==0.4.6
    # via djlint
cssbeautifier==1.15.1
    # via djlint
dj-database-url==2.1.0
django==5.0.6
    # via
    #   dj-database-url
djlint==1.34.1
editorconfig==0.12.4
    # via
    #   cssbeautifier
    #   jsbeautifier
faker==25.2.0
gunicorn==22.0.0
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
jsbeautifier==1.15.1
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.25
    # via djlint
packaging==24.0
    # via gunicorn
pathspec==0.12.1
    # via djlint
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
    # via faker
python-dotenv==1.0.1
pyyaml==6.0.1
    # via djlint
regex==2023.12.25
    # via djlint
ruff==0.4.5
six==1.16.0
    # via
    #   cssbeautifier
    #   jsbeautifier
    #   python-dateutil
sqlparse==0.5.0
    # via django
tqdm==4.66.4
    # via djlint
typing-extensions==4.12.0
    # via dj-database-url
whitenoise==6.6.0
